// app/templates/src/models/IDocument.ts
export interface IDocument {
    id: string;
    title: string;
    url?: string;
    webUrl?: string;
    webTitle?: string;
    previewImageUrl?: string;
    extension?: string;
    created?: string;
    creatorId?: string;
    creatorName?: string;
    creatorPhotoUrl?: string;
    linkToDocument?: string;
}